% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\name{adam}
\alias{adam}
\title{ADAM stochastic process update}
\usage{
adam(k, grad, m, v, beta1 = 0.9, beta2 = 0.999, epsilon = 1e-08)
}
\arguments{
\item{k}{The gradient descent iteration}

\item{grad}{A list of gradient vectors}

\item{m}{A list of momentums}

\item{v}{A list of velocities}

\item{epsilon}{}
}
\value{

}
\description{
ADAM stochastic process update
}
