% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distance_near_row_cpp}
\alias{distance_near_row_cpp}
\title{Calculate thresheld pairwise distance for a row}
\usage{
distance_near_row_cpp(i, locs, locs_grid, radius, byrow = TRUE)
}
\arguments{
\item{i}{The row index}

\item{locs}{An N x 2 matrix of spatial locations}

\item{locs_grid}{An N_grid x 2 matrix of spatial grids}

\item{radius}{The thresholded radius}

\item{byrow}{Perform calculation row/column-wise}
}
\value{
The thresheld pairwise distance
}
\description{
Calculate thresheld pairwise distance for a row
}
