% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nonstationary_MRA.R
\name{fit_nonstationary_MRA}
\alias{fit_nonstationary_MRA}
\title{Title}
\usage{
fit_nonstationary_MRA(
  y,
  locs,
  grid,
  grid_kernels,
  radius = 0.5,
  trunc = 0.01,
  alpha = NULL,
  alpha_a = NULL,
  alpha_b = NULL,
  learn_rate = 0.001,
  rate_schedule = NULL,
  n_iter = 500,
  n_message = 50,
  plot_during_fit = FALSE,
  adam_pars = NULL
)
}
\arguments{
\item{y}{The data}

\item{locs}{An N x 2 matrix of spatial locations}

\item{grid}{A grid object}

\item{grid_kernels}{A grid object for the spatially-varying kernels}

\item{trunc}{The truncation parameter for the kernel parameters}

\item{alpha}{If specified, the top layer MRA parameters}

\item{alpha_a}{If specified, the parameters for the x co-ordinate of the kernels}

\item{alpha_b}{If specified, the parameters for the y co-ordinate of the kernels}

\item{learn_rate}{The gradient descent learning rate}

\item{rate_schedule}{If specified, the gradient descent learning rate schedule in decreasing values.}

\item{n_iter}{The number of gradient descent iterations}

\item{n_message}{The number of iterations between which to output a message}

\item{plot_during_fit}{Plot the current parameter states every \code{n_message} iterations}

\item{adam_pars}{The adam parameter state to allow restarting the model}
}
\description{
Title
}
