% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-deep-MRA-sgd.R
\name{fit_sgd}
\alias{fit_sgd}
\title{Title}
\usage{
fit_sgd(
  y,
  locs,
  grid,
  alpha = NULL,
  alpha_x1 = NULL,
  alpha_y1 = NULL,
  alpha_x2 = NULL,
  alpha_y2 = NULL,
  learn_rate,
  n_iter = 500,
  n_message = 50,
  penalized = FALSE,
  plot_during_fit = FALSE,
  use_spam = FALSE,
  adam_pars = NULL
)
}
\arguments{
\item{use_spam}{}
}
\value{

}
\description{
Title
}
