% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{gradient_fun}
\alias{gradient_fun}
\title{Calculate the Gradient for Regression}
\usage{
gradient_fun(y, tXy, tXX, beta)
}
\arguments{
\item{y}{A vector of observations}

\item{tXy}{A vector of the transpose of design matrix X times the output data y}

\item{tXX}{A matrix of the  transpose of design matrix X times the design matrix X}

\item{beta}{A vector of regression parameters}
}
\value{
The gradient vector
}
\description{
Calculate the Gradient for Regression
}
