% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{gradient_fun_mini}
\alias{gradient_fun_mini}
\title{Calculate the Gradient for Regression using minibatch}
\usage{
gradient_fun_mini(y, tXy, X, tX, beta)
}
\arguments{
\item{y}{A vector of observations}

\item{tXy}{A vector of the transpose of design matrix X times the output data y}

\item{X}{A design matrix}

\item{beta}{A vector of regression parameters}
}
\value{
The gradient vector
}
\description{
Calculate the Gradient for Regression using minibatch
}
