% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{gradient_fun_penalty}
\alias{gradient_fun_penalty}
\title{Calculate the Gradient for Penalized Regression}
\usage{
gradient_fun_penalty(y, tXy, tXX, beta, Q)
}
\arguments{
\item{y}{A vector of observations}

\item{tXy}{A vector of the transpose of design matrix X times the output data y}

\item{tXX}{A matrix of the  transpose of design matrix X times the design matrix X}

\item{beta}{A vector of regression parameters}

\item{Q}{A penalty (precision) matrix}
}
\value{
The gradient vector
}
\description{
Calculate the Gradient for Penalized Regression
}
