% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_basis.R
\name{make_grid}
\alias{make_grid}
\title{Make the deep MRA grid}
\usage{
make_grid(
  locs,
  min_x = -6,
  min_y = -6,
  max_x = 6,
  max_y = 6,
  M = 4,
  n_coarse_grid = 10,
  n_padding = 5L,
  n_neighbors = 68,
  max_points = NULL,
  basis_type = "wendland"
)
}
\arguments{
\item{locs}{An N x 2 matrix of spatial locations}

\item{min_x}{The minimum value of the MRA grid in the x axis}

\item{min_y}{The minimum value of the MRA grid in the y axis}

\item{max_x}{The maximum value of the MRA grid in the x axis}

\item{max_y}{The maximum value of the MRA grid in the y axis}

\item{M}{The number of resolutions.}

\item{n_coarse_grid}{The number of basis functions in one direction (e.g. \code{n_coarse_grid = 10} results in a \eqn{10 \times 10}{10x10} course grid which is further extended by the number of additional padding basis functions given by \code{n_padding}.}

\item{n_padding}{The number of additional boundary points to add on each boundary. For example, n_padding = 5 will add 5 boundary knots to the both the left  and right side of the grid).}

\item{n_neighbors}{The expected number of neighbors for each interior basis function. This determines the basis radius parameter.}

\item{max_points}{The maximum number of points in the}

\item{basis_type}{The basis type. Currently only the "wendland" basis is supported}

\item{use_spam}{Whether to use spam or Matrix for sparse matrix}
}
\value{
A grid object with \code{locs_grid} for the MRA grid and \code{radius} for the thresholded distance radius
}
\description{
Make the deep MRA grid
}
