% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-deep-MRA-sgd.R
\name{predict_deep_MRA}
\alias{predict_deep_MRA}
\title{Title}
\usage{
predict_deep_MRA(
  out,
  grid,
  locs_pred,
  use_spam = FALSE,
  ncores = 1L,
  nchunks = NULL
)
}
\arguments{
\item{out}{The output from fit_sgd}

\item{grid}{The grid}

\item{locs_pred}{The locations at which to predict}

\item{nchunks}{}
}
\description{
Title
}
