% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_gradient_descent.R
\name{regression_gradient_descent}
\alias{regression_gradient_descent}
\title{Fit a regression gradient descent}
\usage{
regression_gradient_descent(
  y,
  X,
  W,
  inits = NULL,
  threshold = 0.01,
  alpha = 0.001,
  num_iters = 10,
  print_every = 1,
  minibatch_size = NULL
)
}
\arguments{
\item{y}{The observed data.}

\item{X}{The design matrix for the fixed effects.}

\item{W}{The basis matrix for the random effects.}

\item{inits}{If NULL, random initialization is used. Otherwise it must
be a numeric vector of length equal to the total columns of X and W.}

\item{threshold}{The threshold for determining convergence.}

\item{alpha}{The learning rate.}

\item{num_iters}{The number of iterations to run the gradient descent algorithm.}

\item{print_every}{The number of iterations between printing the output.}

\item{minibatch_size}{If NULL, the full data are used. If set as a positive numeric value,
this is the number of samples used in the minibatch.}
}
\value{
The output of the gradient descent algorithm
}
\description{
Fit a regression gradient descent
}
