% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_example_splines.R
\name{sim_example_splines}
\alias{sim_example_splines}
\title{Simulate a B-spline mixed model}
\usage{
sim_example_splines(N = 500, df = 8, sigma = 0.5, sigma_alpha = 1.5)
}
\arguments{
\item{N}{The number of observations}

\item{df}{The B-spline degrees of freedom}

\item{sigma}{The residual error standard deviation}

\item{sigma_alpha}{The spline random effect standard deviation}
}
\value{
A list containing the simulated data vector y, the simulated design matrix X,
the simulated vector of fixed effects beta, the simulated B-spline basis Z,
the simulated random effects alpha, and the residual errors epsilon
}
\description{
Simulate a B-spline mixed model
}
