% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{target_fun}
\alias{target_fun}
\title{Calculate the Target function for Regression}
\usage{
target_fun(y, X, beta)
}
\arguments{
\item{y}{A vector of observations}

\item{X}{A design matrix}

\item{beta}{A vector of regression parameters}
}
\value{
The mean square error (MSE) from the linear model
}
\description{
Calculate the Target function for Regression
}
