% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_basis.R
\name{eval_basis}
\alias{eval_basis}
\title{Evaluate the MRA basis}
\usage{
eval_basis(
  locs,
  grid,
  n_neighbors = 68,
  basis_type = "wendland",
  use_spam = TRUE
)
}
\arguments{
\item{locs}{An N x 2 matrix of spatial locations}

\item{grid}{A grid object that is the output of \code{make_grid}}

\item{n_neighbors}{The expected number of neighbors based on the grid}

\item{basis_type}{The basis function type. Currently only "wendland" is accepted}

\item{use_spam}{Whether to use the spam (\code{use_spam = TRUE}) or Matrix (\code{use_spam = FALSE}) package for sparse matrices}
}
\value{

}
\description{
Evaluate the MRA basis
}
