% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_basis.R
\name{eval_basis}
\alias{eval_basis}
\title{Evaluate the MRA basis}
\usage{
eval_basis(
  locs,
  grid,
  n_padding = 5L,
  n_neighbors = 68,
  basis_type = "wendland",
  use_spam = TRUE
)
}
\arguments{
\item{locs}{An N x 2 matrix of spatial locations}

\item{use_spam}{}
}
\value{

}
\description{
Evaluate the MRA basis
}
