% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-deep-MRA-sgd.R
\name{fit_sgd}
\alias{fit_sgd}
\title{Title}
\usage{
fit_sgd(
  y,
  locs,
  grid,
  alpha = NULL,
  alpha_x1 = NULL,
  alpha_y1 = NULL,
  alpha_x2 = NULL,
  alpha_y2 = NULL,
  learn_rate = 0.001,
  rate_schedule = NULL,
  n_iter = 500,
  n_message = 50,
  penalized = FALSE,
  plot_during_fit = FALSE,
  use_spam = FALSE,
  adam_pars = NULL,
  sparse_outer = TRUE,
  noisy = TRUE
)
}
\arguments{
\item{y}{The data}

\item{locs}{An N x 2 matrix of spatial locations}

\item{grid}{A grid object that is the output of \code{make_grid}}

\item{alpha}{If specified, the top layer MRA parameters}

\item{alpha_x1}{If specified, the first hidden layer MRA parameters for the x-axis variable}

\item{alpha_y1}{If specified, the first hidden layer MRA parameters for the y-axis variable}

\item{alpha_x2}{If specified, the second hidden layer MRA parameters for the x-axis variable}

\item{alpha_y2}{If specified, the second hidden layer MRA parameters for the y-axis variable}

\item{learn_rate}{The gradient descent learning rate}

\item{rate_schedule}{If specified, the gradient descent learning rate schedule in decreasing values.}

\item{n_iter}{The number of gradient descent iterations}

\item{n_message}{The number of iterations between which to output a message}

\item{penalized}{Fit using a penalty term}

\item{plot_during_fit}{Plot the current parameter states every \code{n_message} iterations}

\item{use_spam}{Whether to use the spam (\code{use_spam = TRUE}) or Matrix (\code{use_spam = FALSE}) package for sparse matrices}

\item{adam_pars}{The adam parameter state to allow restarting the model}

\item{sparse_outer}{If \code{TRUE}, calculate the outer product in a sparse format. For all but the smallest models, this should be TRUE. I should make this automatic going forward}

\item{noisy}{If \code{TRUE}, add random noise to the gradient.}
}
\value{

}
\description{
Title
}
