% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-deep-mra.R
\name{sim_deep_mra}
\alias{sim_deep_mra}
\title{Simulate deep MRA}
\usage{
sim_deep_mra(
  N = 100^2,
  M = 1,
  n_coarse_grid = 20,
  n_layers = 3,
  sigma = 0.25,
  use_spam = FALSE,
  ncores = 1L,
  nchunks = NULL
)
}
\arguments{
\item{N}{The number of locations}

\item{M}{The number of resolutions in the MRA approximation to the Gaussian process}

\item{n_coarse_grid}{The number of grid locations in the coarsest resolution of the MRA approximation to the Gaussian process}

\item{n_layers}{The number of layers in the deep process}

\item{sigma}{The observation standard deviation}

\item{use_spam}{A logical value of whether to use the spam library (use_spam=TRUE) or the Matrix library (use_spam=FALSE) for sparse matrix calculations}

\item{ncores}{The number of cores to use for parallelization}

\item{nchunks}{The number of chunks to divide the distance calculation into. The default argument of NULL will use the same number of chunks as the number of cores.}
}
\description{
Simulate deep MRA
}
